Refactoring library module a bit.
