+++ library/src/main/java/com/adammcneilly/tournament/bracket/internal/BracketMatchItem.kt
@@ -1,22 +1,16 @@
-package com.adammcneilly.tournament.bracket
+package com.adammcneilly.tournament.bracket.internal
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Divider
 import androidx.compose.material3.LocalContentColor
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
+import com.adammcneilly.tournament.bracket.displaymodels.BracketMatchDisplayModel
 
 /**
  * Defines an item that is used to represent a [match] inside of a bracket.
@@ -54,42 +48,5 @@ internal fun BracketMatchItem(
                 color = contentColor,
             )
         }
-    }
-}
-
-@Composable
-private fun TeamRow(
-    team: BracketTeamDisplayModel,
-    color: Color,
-) {
-    val fontWeight: FontWeight? = if (team.isWinner) {
-        FontWeight.Bold
-    } else {
-        null
-    }
-
-    Row {
-        Text(
-            text = team.name,
-            fontWeight = fontWeight,
-            color = color,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier
-                .padding(8.dp)
-                .weight(1F),
-        )
-
-        Spacer(modifier = Modifier.width(8.dp))
-
-        Text(
-            text = team.score,
-            fontWeight = fontWeight,
-            color = color,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier
-                .padding(8.dp),
-        )
     }
 }
