+++ library/src/main/java/com/adammcneilly/tournament/bracket/internal/TeamRow.kt
@@ -1,0 +1,54 @@
+package com.adammcneilly.tournament.bracket.internal
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import com.adammcneilly.tournament.bracket.displaymodels.BracketTeamDisplayModel
+
+/**
+ * Component to render a [team] inside of a [BracketMatchItem].
+ */
+@Composable
+internal fun TeamRow(
+    team: BracketTeamDisplayModel,
+    color: Color,
+) {
+    val fontWeight: FontWeight? = if (team.isWinner) {
+        FontWeight.Bold
+    } else {
+        null
+    }
+
+    Row {
+        Text(
+            text = team.name,
+            fontWeight = fontWeight,
+            color = color,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis,
+            modifier = Modifier
+                .padding(8.dp)
+                .weight(1F),
+        )
+
+        Spacer(modifier = Modifier.width(8.dp))
+
+        Text(
+            text = team.score,
+            fontWeight = fontWeight,
+            color = color,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis,
+            modifier = Modifier
+                .padding(8.dp),
+        )
+    }
+}
